Hacer refactoring apartir de este commit.. Danny!!!
