+++ PublicTransport/src/DB/Android/ServerToAndroid.java
@@ -2,7 +2,6 @@ package DB.Android;
 
 import java.net.ServerSocket;
 import java.net.Socket;
-import DateBase.SocketClientThread;
 
 public class ServerToAndroid {
     public static void main(String[] args) {
