+++ PublicTransport/src/DB/Android/SocketClientThread.java
@@ -1,4 +1,4 @@
-package DateBase;
+package DB.Android;
 
 import java.io.ObjectInputStream;
 import java.io.ObjectOutputStream;
